Fix for building on OmniOS

+++ cop/TrafficCop.cc
@@ -249,7 +249,7 @@ sig_fatal(int signum)
   if (t) {
     if (t->si_code <= 0) {
       cop_log(COP_FATAL, "cop received fatal user signal [%d] from"
-              " pid [%d] uid [%d]\n", signum, t->si_pid, t->si_uid);
+              " pid [%d] uid [%d]\n", signum, (int)(t->si_pid), t->si_uid);
     } else {
       cop_log(COP_FATAL, "cop received fatal kernel signal [%d], " "reason [%d]\n", signum, t->si_code);
     }
